backend/requirements.txt
fastapi==0.99.0
uvicorn[standard]==0.23.0
sqlmodel==0.0.8
SQLAlchemy==2.2.4
asyncpg==0.28.0
python-dotenv==1.0.0
openai==1.29.0
httpx==0.24.1
python-multipart==0.0.6
pytest==7.4.2
python-jose[cryptography]==3.3.0
passlib==1.9.0
argon2-cffi==21.3.0
alembic==1.11.1
sendgrid==6.11.0
jinja2==3.1.2
slowapi==0.1.5
sentry-sdk==1.29.0
prometheus-fastapi-instrumentator==6.0.1
redis==4.5.5
